\name{equate-package}
\alias{equate-package}
\docType{package}
\title{
Observed-Score Linking and Equating
}
\description{
Contains methods for observed-score linking and equating under the single-group, equivalent-groups, and nonequivalent-groups with anchor test(s) designs. Equating types include identity, mean, linear, general linear, equipercentile, circle-arc, and composites of these. Equating methods include synthetic, nominal weights, Tucker, Levine observed score, Levine true score, Braun/Holland, frequency estimation, and chained equating. Plotting and summary methods, and methods for multivariate presmoothing and bootstrap error estimation are also provided.}
\details{
\tabular{ll}{
Package: \tab equate\cr
Version: \tab 2.0-4\cr
Date: \tab 2015-10-21\cr
Depends: \tab R (>= 3.0.0)\cr
License: \tab GPL (>= 2)\cr
}

Index:
\preformatted{
ACTmath                 ACT Mathematics Test Scores
KBneat                  Test Scores under a NEAT design
PISA                    Programme for International Student Assessment 2009 USA Data
descriptives            Descriptive Statistics for Frequency Tables
percentiles             Percentile ranks and Cumulative Frequencies for Frequency Tables
equate                  Observed Score Linking and Equating
composite               Composite Linking and Equating
bootstrap               Bootstrap Equating Error
freqtab                 Frequency Distribution Tables
presmoothing            Frequency Distribution Presmoothing
plot.freqtab            Plotting Frequency Distributions
plot.equate             Plotting Equating Results
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{equatevignette} \tab equate vignette (source, pdf)\cr
}
The package vignette provides an introduction to linking and equating and includes descriptions of the supported equating methods and examples. The help page for the main function of the package, \code{\link{equate}}, contains additional examples.
}
\author{
Anthony Albano <tony.d.albano@gmail.com>
}
\keyword{package}
