\name{print.equate}
\Rdversion{1.1}
\alias{print.equate}
\title{Print Method for Class 'equate'}
\description{
This function extracts and prints default output for objects of class \dQuote{\code{\link{equate}}}.
}
\usage{
print.equate(x, ...)
}
\arguments{
  \item{eq}{
object of class \dQuote{\code{\link{equate}}}
}
  \item{\dots}{
further arguments passed to or from other methods
}
}
\author{
Anthony Albano \email{tony.d.albano@gmail.com}
}
\examples{
}
\keyword{internal}